+++ src/sphinx/changelog.rst
@@ -3,7 +3,7 @@ Changelog
 
 Changes in Version 2.6.X
 ------------------------
-
+- Publish casbah.commons test helpers
 - Added suport $setOnInsert to the query dsl
 
 Changes in Version 2.6.0